sna: Apply defense against a rogue call to move a SW cursor
So the sna_set_cursor_postion() assertion was reported to have caught us
trying to adjust the position of an absent cursor. That should be
impossible as we should only arrive there if we claim we can support a
HW cursor and so have a registered cursor. However, changing the
assertion into a guard is trivial, and preserves peace of mind.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>