sna: Do not discard damage when using the CPU to perform BLT spans
The BLT paths are only taken when we know we are overwriting the target
contents, and so we do not have to worry about reads along those paths.
However, in terms of migrating the damage for a pixmap, we have to be
careful in case we do not write to the full area of the composite
rectangle, in which case we have to treat it as a read to that area and
migrate the damage.

Regression from
commit a13781d19defc97af6a279c11a85e33ef825020e [2.20.10]
Author: Chris Wilson <chris@chris-wilson.co.uk>
Date:   Sun Aug 19 09:45:12 2012 +0100

    sna: Enable BLT composite functions to target CPU buffers

    Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>

Reported-by: Matti Hämäläinen
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>