sna: Prefer PrimaryOutput for timing signals
If the drawable is even partially on the PrimaryOutput, use that pipe
for determining synchronisation counters and synchronised updates.

OML_sync_control:
	"For a multi-monitor system, the monitor used to
	 determine MSC is screen 0 of <display>."

The PrimaryOutput is the nearest approximation we have to "screen 0". We
also deviate in handling what hapens when that screen is off, which is
not mentioned at all in OML_sync_control.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>