sna/dri2: Adjust current msc / target seq between CRTCs
The target_msc supplied by the client is required to be monotonic. Since
the client does not know which CRTC the window is on, they do not
control from which pipe we derive the MSC, and so we need to adjust the
hardware values such that the client only ever sees a monotonic value.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>