Correct clearing for overscan
When we add overscan borders we have to make sure that they do not show
stale data, since the borders might actually be (partially) visible.

To do this, we were clearing display buffers during modesetting, but
that turned out to be too aggressive, clearing buffers (which were not
going to be repainted) when restoring from DPMS off.

Move the buffer clearing code to a more appropriate place.