sna: Disable rendering with the DRM device whilst away from VT
As root, X gets away with many things, including submitting commands to
the DRM device whilst it is no longer authorised (i.e. when it has
relinquished master to another client across a VT switch). In the
non-root future, if we attempt to use the device whilst unauthorized the
rendering will be lost and we will mark the device as unusable. So flush
our render queue to the device around a VT switch.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>