sna: Keep the backlight at the same brightness when switching away
Before VT switching, we disable all the outputs. This has the
side-effect of disabling the backlight, but commonly the console does
not set the backlight brightness itself and so we re-enable the
backlight on its behalf. Rather than set the backlight to max
brightness, reset it to the last user value, as maximum can be very
bright!

Reported-by: Mark Kettenis <mark.kettenis@xs4all.nl
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>