sna: Treat external bo carefully when marking as idle
When we flag a bo as being externally modified, we set a special flag on
its request pointer. This forces us to inspect whether the bo is busy
before accessing it the next, and if it is idle, we must be careful that
we then do not dereference the special request.

Fixes regression from
commit 9115406b651eb11779f37c4b048d55188b9d6e02
Author: Chris Wilson <chris@chris-wilson.co.uk>
Date:   Thu Oct 30 19:21:02 2014 +0000

    sna: Retire all requests when a bo is found to be not busy

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>