sna: Skip allocation of new TearFree buffer if not actually flipping
After showing the new front buffer, we have to avoid writing to the back
buffer whilst the flip is still pending (as it may still be being shown
by the display). To do this we check if there are any outstanding flips,
and reallocate if so. However, with a CRTC override, we may just be
flipping that one CRTC and not the current TearFree back buffer and in
that case we can avoid the reallocation.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>