sna/gen7: Limit threads on HSW GT1
For whatever reason, it seems that for HSW GT1 we cannot specify the
maximum value of the field and leave it to the hardware to clamp the
value to the maximum supported. The impact should be zero other than the
possibilty it workarounds an issue if the hardware doesn't apply the
limit.

References: https://bugs.freedesktop.org/show_bug.cgi?id=87564
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>