sna/dri2: Only preserve back buffers with the same pitch
We can only pageflip if we don't change the pitch, otherwise we
simulate the pageflip with modeset. The idea is that once we have
exchanged the buffers, we will quickly settle down to a steady state
scenario where each buffer has the same pitch. We can encourage this
further by only keeping those buffers around that do have the same pitch
as the current front.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>