sna/present: Restrict pending flip check to enqueueing
Present checks for a flip-compatible Pixmap prior to deciding whether to
queue the flip after a pending flip, and at that point we do not want to
reject a potential flip due to outstanding flips. That is only of
restriction only applies when we request the kernel to perform the flip,
so more the check to the corresponding position.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>