sna/present: Avoid tricks with the TearFree flag
Modifying the flag may cause havoc if we need to modeset whilst Present
is in control as the CRTC will be setup without reference to TearFree, but
then when Present unflips, we restore the stale configuration. So let's
have a separate flag just to disable TearFree flips when a Present flip
chain is active.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>