sna: Flush BLT operations to an idle GPU
We improve dispatch latency if after creating a command buffer we
immediately submit if the GPU is idle. This improves concurrency as we
continue to build the next command buffer as the GPU executes and helps
prevent needlessly using one engine for too long (i.e. sometimes we may
be able to execute the work much early and do the ring switch cheaply).

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>