sna: Enable use of GPU tiling even when SET_TILING fails
On gen4+, we no longer need to setup a fence for tiled operations and so
can program the GPU irrespective of the kernel tiled settings. However,
this means that we then cannot access the object through a GTT mmapping
(as we have no fence to do detiling) and nor can we use a WC mmap as the
swizzle is unknown.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>