sna: Pass scanout flag when creating PRIME bo
For PRIME bo, we need to use uncached render targets so that any writes
are flushed out to main memory where they can be immediately read by a
PCI device. For simplicity, we just request that PRIME bo be also
SCANOUTs as that ensures that they will be created with the right
attributes for coherent main memory.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>