sna/dri2: Mark the pending backbuffer copy as active
In order to perform swap elision for windows, we defer the blit until
the next vblank and then do whatever was the last buffer to be presented
by the client. Whilst that copy is pending, in the same manner as a
pending flip, we cannot hand that buffer back to the client for use (or
else they will render over top of it before we copy from it, or even
worse we copy from it in the middle of rendering).

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>