sna/dri2: SWAP_THROTTLE is also used for a delayed SWAP
If there is already a pending operation on the drawable, the swap is
queued. It is queued with SWAP_THROTTLE which classes with the
vblank_mode=0 chaining, so the choice is to either review all other uses
of SWAP_THROTTLE and duplicate the chain swap handling, or to choose when
use to the alternate copy or back buffer. This patch opts for the
latter.

Fixes regression from
commit dcb4d323ca19f86fbe0230378ac9035161a70f9e
Author: Chris Wilson <chris@chris-wilson.co.uk>
Date:   Thu Jun 11 13:54:49 2015 +0100

    sna/dri2: Mark the pending backbuffer copy as active

Reported-by: Christoph Haag <haagch@frickel.club>
Reported-by: Andreas Reis <andreas.reis@gmail.com>
Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=90968
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>